requirements.txt
streamlit
pandas
plotly
PyYAML
psycopg2-binary
sqlalchemy
python-dotenv

# A retirar futuramente
openpyxl>=3.1.0